move getProductsList to handlers folder
